requirements.txt
flask
pinecone-client
snowflake-connector-python
scikit-learn
numpy
requests
nltk
gunicorn
flask-cors
